+++ hyperspy/_components/eels_arctan.py
@@ -52,7 +52,7 @@ class EELSArctan(Expression):
     """
 
     def __init__(self, A=1., k=1., x0=1., module=["numpy", "scipy"], **kwargs):
-        # Not to break scripts once we remove the legacy Arctan
+        # To be able to still read old file versions that contain this argument
         if "minimum_at_zero" in kwargs:
             del kwargs["minimum_at_zero"]
         super().__init__(