+++ hyperspy/model.py
@@ -113,10 +113,12 @@ def reconstruct_component(comp_dictionary, **init_args):
         # in HyperSpy 1.6 the old Voigt component was moved to PESVoigt
         if comp_dictionary['parameters'][4]['_id_name'] == "resolution":
             comp_dictionary['_id_name'] = "PESVoigt"
+            _logger.info("Legacy Voigt component converted to PESVoigt during file reading.")
     if (comp_dictionary['_id_name'] == "Arctan" and 'minimum_at_zero' in comp_dictionary):
         # in HyperSpy 1.6 the old Arctan component was moved to EELSArctan
         if comp_dictionary['minimum_at_zero'] == True:
             comp_dictionary['_id_name'] = "EELSArctan"
+            _logger.info("Legacy Arctan component converted to EELSArctan during file reading.")
     _id = comp_dictionary['_id_name']
     if _id in _COMPONENTS:
         _class = getattr(