+++ hyperspy/tests/component/test_arctan.py
@@ -32,10 +32,3 @@ def test_function():
     np.testing.assert_allclose(g.function(1), 0)
     np.testing.assert_allclose(g.function(1e4), 10*np.pi/2,1e-4)
 
-def test_function_delattr():
-    g = Arctan(minimum_at_zero=True, legacy=True)
-    g.A.value = 10
-    g.k.value = 2
-    g.x0.value = 1
-    np.testing.assert_allclose(g.function(0), -11.07148718)
-    np.testing.assert_allclose(g.function(1), 0)